requirements.txt
numpy==1.24.1
pandas==1.4.2
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.17.0
tensorflow==2.11.0
